+++ src/init.c
@@ -8,6 +8,9 @@
 */
 
 /* .Call calls */
+extern SEXP C_character2integer(SEXP, SEXP, SEXP, SEXP);
+extern SEXP C_comma(SEXP, SEXP, SEXP);
+extern SEXP C_empty(SEXP, SEXP, SEXP, SEXP);
 extern SEXP C_hausdorffEuclid(SEXP, SEXP);
 extern SEXP C_haversineDistance(SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP C_match_min_Haversine(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
@@ -91,6 +94,9 @@ extern SEXP Cwhich3_mem(SEXP, SEXP, SEXP, SEXP);
 extern SEXP fmatch(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 
 static const R_CallMethodDef CallEntries[] = {
+    {"C_character2integer",           (DL_FUNC) &C_character2integer,            4},
+    {"C_comma",                       (DL_FUNC) &C_comma,                        3},
+    {"C_empty",                       (DL_FUNC) &C_empty,                        4},
     {"C_hausdorffEuclid",             (DL_FUNC) &C_hausdorffEuclid,              2},
     {"C_haversineDistance",           (DL_FUNC) &C_haversineDistance,            5},
     {"C_match_min_Haversine",         (DL_FUNC) &C_match_min_Haversine,         11},